🚩(endpoints): add new pausing capabilities
Includes two new functions on endpoints - `pause` and `resume`.
`pause` take a single argument, `cold`, which manages whether or not to still handle any messages currently being processed.

To manage this functionality, a few promises are passed around (`this._starting`, `this._resuming`, `this._paused` and `this._cold`).
This is currently necessary to fight off multiple calls to the same functions and to correctly handle interrupting processing if a cold pause has been requested.

This will also need adding to listeners as well as docs and tests adding.