🧪 include EAI_AGAIN in connection test
CI that can't access the outside world fails here and EAI_AGAIN only
shows a local network issue anyway.

The aim of this test is to show it fails predictably when trying to
connect to something that doesn't exist, so that still holds true if
we have multiple accepted error codes.